Set build targets to not build for testing when unchanged
After https://github.com/square/refinement/pull/8, when filtering for
testing, we leave build targets unchanged, so as not to break macro
expansion for the test targets that don't get removed.

This caused an unfortunate side-effect of (unnecessarily) building the
build target unconditionally when running
`xcodebuild build-for-testing`.

By setting `buildForTesting = "NO"` in the scheme when the build target
is unchanged, it will still be available for macro expansion,
but it will not be built when running `xcodebuild build-for-testing`.